added edit property
